+++ R/sfc.R
@@ -144,7 +144,8 @@ sfc_linestring <- function( obj = NULL, x = NULL, y = NULL, z = NULL, m = NULL, 
 #'
 #'
 #' @export
-sfc_multilinestring <- function( obj = NULL, x = NULL, y = NULL, z = NULL, m = NULL, multilinestring_id = NULL, linestring_id = NULL ) {
+sfc_multilinestring <- function( obj = NULL, x = NULL, y = NULL, z = NULL, m = NULL,
+                                 multilinestring_id = NULL, linestring_id = NULL ) {
   geometry_columns <- c(x,y,z,m)
   rcpp_sfc_multilinestring( obj, index_correct( geometry_columns ), index_correct( multilinestring_id ), index_correct( linestring_id ) )
 }
@@ -203,7 +204,8 @@ sfc_multilinestring <- function( obj = NULL, x = NULL, y = NULL, z = NULL, m = N
 #'
 #'
 #' @export
-sfc_polygon <- function( obj = NULL, x = NULL, y = NULL, z = NULL, m = NULL, polygon_id = NULL, linestring_id = NULL, close = TRUE ) {
+sfc_polygon <- function( obj = NULL, x = NULL, y = NULL, z = NULL, m = NULL,
+                         polygon_id = NULL, linestring_id = NULL, close = TRUE ) {
   geometry_columns <- c(x,y,z,m)
   rcpp_sfc_polygon( obj, index_correct( geometry_columns ), index_correct( polygon_id ), index_correct( linestring_id ), close )
 }
@@ -283,7 +285,51 @@ sfc_polygon <- function( obj = NULL, x = NULL, y = NULL, z = NULL, m = NULL, pol
 #' sfc_multipolygon( df, x = "x", y = "y", polygon_id = "id1")
 #'
 #' @export
-sfc_multipolygon <- function( obj = NULL, x = NULL, y = NULL, z = NULL, m = NULL, multipolygon_id = NULL, polygon_id = NULL, linestring_id = NULL, close = TRUE ) {
+sfc_multipolygon <- function( obj = NULL, x = NULL, y = NULL, z = NULL, m = NULL,
+                              multipolygon_id = NULL, polygon_id = NULL, linestring_id = NULL,
+                              close = TRUE ) {
   geometry_columns <- c(x,y,z,m)
   rcpp_sfc_multipolygon( obj, index_correct( geometry_columns ), index_correct( multipolygon_id ), index_correct( polygon_id ), index_correct( linestring_id ), close )
+}
+
+
+#' Calculate bbox
+#'
+#' Calculates the bounding box of a \code{sfc} object. This does not read the "bbox" attribute,
+#' it re-calculates the bounding box from the geometry coordinates
+#'
+#' @params sfc simple-feature collection \code{sfc} object.
+#'
+#' @examples
+#'
+#' df <- data.frame(
+#' id1 = c(1,1,1,1,1,1,1,1,2,2,2,2)
+#' , id2 = c(1,1,1,1,2,2,2,2,1,1,1,1)
+#' , x = c(0,0,1,1,1,1,2,2,3,4,4,3)
+#' , y = c(0,1,1,0,1,2,2,1,3,3,4,4)
+#' )
+#'
+#' pt <- sfc_point(obj = df, x = "x", y = "y", z = "id1")
+#' mpt <- sfc_multipoint(obj = df, x = "x", y = "y", multipoint_id = "id1")
+#' ls <- sfc_linestring(obj = df, x = "x", y = "y", linestring_id = "id1")
+#' mls <- sfc_multilinestring(obj = df, x = "x", y = "y", multilinestring_id = "id1")
+#' p <- sfc_polygon(obj = df, x = "x", y = "y", polygon_id = "id1", linestring_id = "id2", close = FALSE )
+#' mp <- sfc_multipolygon(obj = df, x = "x", y = "y", multipolygon_id = "id1", linestring_id = "id2", close = FALSE )
+#'
+#' calculate_bbox( pt )
+#' calculate_bbox( mpt )
+#' calculate_bbox( ls )
+#' calculate_bbox( mls )
+#' calculate_bbox( p )
+#' calculate_bbox( mp )
+#'
+#' ## you can use it to update a bounding-box if it gets corrupted
+#' attr( mpt, "bbox" ) <- c(1:5)
+#' mpt ## incorrect values
+#' attr( mpt, "bbox" ) <- calculate_bbox( mpt )
+#' mpt ## back to correct values
+#'
+#' @export
+calculate_bbox <- function( sfc ) {
+  return( rcpp_get_bbox( sfc ) )
 }